Specify project file in getCapabilites url
The instructions did not work because a regression caused by ea22184,
which removed the QGIS_PROJECT_FILE environment variable. Fixed by providing
the MAP parameter pointing to the project file in the url. Also use qgz format
instead of qgs since it is the default format when saving in QGIS Desktop.